+++ web/documentserver-example/csharp/DocEditor.aspx.cs
@@ -140,9 +140,8 @@ namespace OnlineEditorsExample
 
             // get file extension
             var ext = Path.GetExtension(FileName).ToLower();
-            var canFill = _Default.FillFormsExts.Contains(ext);
             // get editor mode or set the default one (edit)
-            var editorsMode = Request.GetOrDefault("editorsMode", canFill ? "fillForms" : "edit");
+            var editorsMode = Request.GetOrDefault("editorsMode", "edit");
 
             var canEdit = _Default.EditedExts.Contains(ext);  // check if this file can be edited
             var editorsType = Request.GetOrDefault("editorsType", "desktop");
@@ -150,7 +149,7 @@ namespace OnlineEditorsExample
             var id = Request.Cookies.GetOrDefault("uid", null);
             var user = Users.getUser(id);  // get the user
             
-            if ((!canEdit && editorsMode.Equals("edit") || editorsMode.Equals("fillForms")) && canFill) {
+            if ((!canEdit && editorsMode.Equals("edit") || editorsMode.Equals("fillForms")) && _Default.FillFormsExts.Contains(ext)) {
                 editorsMode = "fillForms";
                 canEdit = true;
             }